// reference-code/esr/esale-mobile/src/shared/components/suggestions/interface/task-suggest-interface.ts
export interface TaskSuggest {
  taskId: number,// data in database mapping data form
  parentCustomerName: string, // data in database mapping
  milestoneName: string,// data in database mapping
  customerName: string,// data in database mapping
  productTradingName: string,
  taskName: string,// data in database mapping
  finishDate: string,// data in database mapping
  finishType: number,// data in database mapping
  employeeName: string,// data in database mapping
  status: number,// data in database mapping
  data: any
}

/**
 * Define milestone suggest view
 */
export interface IResultSearchProps {
  typeSearch: number, // type search (SINGLE or MULTI)
  searchConditions: ISearchCondition[],
  isRelation?: boolean,
  updateStateElement: (searchValue: any) => void; // callback when change status control
  closeModal: () => void;
  exportError: (err: any) => void;
}
// ...
export interface ITask {
  taskId: number,
  taskName: string,
  statusTaskId: number,
  finishDate: string,
  milestoneId: number,
  milestoneName: string,
  customers: {
    customerId: number,
    customerName: string,
    parentCustomerName: string
  }[],
  productTradings: {
    productTradingId: number,
    productName: string
  }[],
  employees: {
    employeeId: number,
    employeeName: string
  }[]
}